% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoSearch.R
\name{geoSearch}
\alias{geoSearch}
\title{Search locations by name}
\usage{
geoSearch(search, apiKey)
}
\arguments{
\item{search}{Search term (required).}

\item{apiKey}{Your API key (required).}
}
\description{
Use this function to search for locations within the AskCHIS NE database
by name and obtain geoId, and location type.
}
\examples{
geoSearch(search = 'Los Angeles', apiKey = '<YOUR API KEY>')
Returns a data frame with all locations that match 'Los Angeles.'
}
\keyword{askchis}
\keyword{chis}

