% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{Get AskCHIS NE Metadata}
\usage{
getMetadata(apiKey)
}
\arguments{
\item{apiKey}{Your API key (required).}
}
\description{
This function obtains all of the metadata available in AskCHIS NE. This includes
indicator labels, categories, variable years etc.
}
\examples{
getMetadata(apiKey = '<YOUR API KEY>')
Returns a data frame with all of the metadata in AskCHIS NE.
}
\keyword{askchis}
\keyword{chis}

