% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolEstimate.R
\name{poolEstimate}
\alias{poolEstimate}
\title{Pool AskCHIS NE Estimates}
\usage{
poolEstimate(indicator, attributes = NULL, locations, apiKey)
}
\arguments{
\item{indicator}{The indicator id (can be obtained using getMetadata()) (required).}

\item{attributes}{The specific attributes requested. A list that can include: estimate, population, SE, CI_L95, CI_U95, CV, MSE. Will return all if not specified (optional).}

\item{locations}{The specific locations requested to be pooled. A list that must include geoIds (can be obtained from geoSearch()) (required).}

\item{apiKey}{Your API key (required).}
}
\description{
This function combines estimates for multiple geographic locations and is especially
useful in combining multiple locations that might have suppressed estimates to get an overall estimate.
}
\examples{
poolEstimate(indicator = 'OBESEA', attributes = list('estimate','population'), locations = list('666000','644000'), apiKey = <YOUR API KEY>)
Returns a data frame with adult obesity estimates and populations for Los Angeles and San Diego cities COMBINED.
}
\keyword{askchis}
\keyword{chis}

