% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setKey.R
\name{setKey}
\alias{setKey}
\title{Set-up AskCHIS API Key}
\usage{
setKey(apiKey)
}
\arguments{
\item{apiKey}{Your API key (required).}
}
\description{
All functions require an API key. This function allows the user to define their API key once per session
rather than explicitely passing the API key with each query.
}
\examples{
\dontrun{
setKey('<YOUR API KEY>')
}
}
\keyword{api}
\keyword{askchis}
\keyword{chis}
\keyword{key}

